CellosticsApplication/requirements.txt
Sphinx
Django==1.9.7
django-filter==0.11.0
djangorestframework==3.3.3
drf-nested-routers==0.10.0
psycopg2
django-debug-toolbar
django-cors-headers
django-extensions
werkzeug
reportlab
Werkzeug
#pip install --pre SQLAlchemy --upgrade
#zypper install rabbitmq-server   # For starting RabbitMQ 'rabbitmq-server start'
pip install redis  # For starting Redis 'redis-server'
# Install RabbitMQ before celery
#librabbitmq
celery   # export C_FORCE_ROOT="true"   (If you want to run it from root)
django-celery
pika

